修复构建错误并完成 CodebaseMemory 集成
- CodebaseMemoryRunner: 移除 @MainActor 避免 init 跨 actor 调用
- CodebaseMemoryViewModel: 移除 EntitlementGate 依赖(Pro 门控已在 RepoListView), 修复 Self.emptySteps() stored property
- CodebaseMemoryStorage: 修复 Int→Int64 转换
- CodebaseMemoryPanel: 修复 footerStatus String→Text
- IntegrationSettingsView: 修复 Section 插入位置(Form body 内)
- Localizable.xcstrings: Python 脚本整体重建所有键(JSON 合法)
- 删除重复 UPSTREAM-README.md(与 CodeFlow 同名冲突)

BUILD SUCCEEDED

+++ Starcat/Features/CodeGraph/CodebaseMemory/CodebaseMemoryStorage.swift
@@ -141,7 +141,7 @@ struct CodebaseMemoryStoredProject: Identifiable, Equatable, Sendable {
 
     /// 递归目录总大小（字节）。metadata.json + source/ + .codebase-memory/ 等全部计入。
     var totalBytes: Int64 {
-        (try? Self.directorySize(of: directoryURL)) ?? 0
+        Int64((try? Self.directorySize(of: directoryURL)) ?? 0)
     }
 
     /// 最近活跃时间（UI 排序用）。
